fixed repeated api call issue
