+++ src/components/IssueDescription.js
@@ -58,7 +58,7 @@ function IssueDescription({ currentOrg, currentIssue }) {
                   <li className="breadcrumbs__item">
                     <a
                       className="breadcrumbs__link breadcrumbs-items"
-                      href={`/docs/2024/org?id=${currentOrg}`}
+                      href={`/docs/2024?id=${currentOrg}`}
                     >
                       {currentOrg}
                     </a>
@@ -330,7 +330,7 @@ function IssueDescription({ currentOrg, currentIssue }) {
                           {d.content && isLearningsDropdown == i && (
                             <tr style={{ backgroundColor: "none" }}>
                               <td align="left">
-                                <Markdown>{d.content}</Markdown>
+                                {d.content.match(/<ul>[\s\S]*?<\/ul>/) ? <Markdown>{d.content.match(/<ul>[\s\S]*?<\/ul>/)[0]}</Markdown> : <Markdown>{d.content}</Markdown>}
                               </td>
                             </tr>
                           )}
